requirements.txt
deepgram-sdk
httpx
pyaudio
pvporcupine
webrtcvad
requests
python-dotenv
elevenlabs
fastapi
uvicorn
